// frontend/src/types/openapi.d.ts
export interface components {
    schemas: {
// ...
        RefundOut: {
            /** Acknowledge Used */
            acknowledge_used: boolean;
            /** Amount */
            amount: string;
            /**
             * Club Id
             * Format: uuid
             */
            club_id: string;
            /** Completed At */
            completed_at?: string | null;
            /** Completed By User Id */
            completed_by_user_id?: string | null;
            /** Currency */
            currency: string;
            /** Decided At */
            decided_at?: string | null;
            /** Decided By User Id */
            decided_by_user_id?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Manual Reference */
            manual_reference?: string | null;
            /**
             * Order Id
             * Format: uuid
             */
            order_id: string;
            /**
             * Payment Id
             * Format: uuid
             */
            payment_id: string;
            /** Physical Return */
            physical_return: boolean;
            /** Reason */
            reason: string;
            /**
             * Requested At
             * Format: date-time
             */
            requested_at: string;
            /**
             * Requested By User Id
             * Format: uuid
             */
            requested_by_user_id: string;
            /** Status */
            status: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}